requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
openai>=1.0.0
gspread==5.7.2
oauth2client==4.1.3
google-api-python-client==2.99.0
elevenlabslib==0.32.0
requests==2.31.0